moving racks to own tab
