+++ src/components/elementView/elementTabContainer.tsx
@@ -6,11 +6,14 @@ import { RouteComponentProps } from "react-router";
 import "./elementView.scss";
 import { connect } from "react-redux";
 import { ElementType } from "../../utils/utils";
+import RackTab from "./rackTab";
 
 interface ElementTabContainerProps {
   isAdmin: boolean;
 }
-class ElementTabContainer extends React.Component<ElementTabContainerProps & RouteComponentProps> {
+class ElementTabContainer extends React.Component<
+  ElementTabContainerProps & RouteComponentProps
+> {
   public render() {
     return (
       <Tabs
@@ -19,7 +22,7 @@ class ElementTabContainer extends React.Component<ElementTabContainerProps & Rou
         id="ElementViewer"
         key={"vertical"}
         renderActiveTabPanelOnly={false}
-        vertical={false}
+        vertical={true}
         large
       >
         <Tab
@@ -34,12 +37,8 @@ class ElementTabContainer extends React.Component<ElementTabContainerProps & Rou
           title="Models"
           panel={<ElementTab {...this.props} element={ElementType.MODEL} />}
         />
-        {/* <Tab
-          className="tab"
-          id="rack"
-          title="Racks"
-          panel={<ElementView element="racks" />}
-        /> */}
+        <Tab className="tab" id="rack" title="Racks" panel={<RackTab />} />
+        <Tab className="tab" id="datacenter" title="Datacenters" />
         <Tabs.Expander />
       </Tabs>
     );
