+++ src/components/elementView/elementView.scss
@@ -2,6 +2,11 @@
   display: flex;
   flex-direction: row;
 }
+
+.tab {
+  width: 100%;
+  padding: 10pt;
+}
 .tab:focus {
   outline: none;
 }
